Broke out models per https://angular.io/guide/styleguide#style-01-01
