+++ src/app/shared/models/index.ts
@@ -1,0 +1,3 @@
+export * from './lesson.model';
+export * from './quiz-answer.model';
+export * from './topic.model';
