+++ src/app/shared/models/lesson.model.ts
@@ -1,0 +1,5 @@
+import { TopicModel } from '.';
+
+export interface LessonModel extends TopicModel {
+  quizId: number;
+}
