+++ src/app/shared/models/quiz-answer.model.ts
@@ -1,0 +1,6 @@
+export interface QuizAnswerModel {
+  id: string;
+  givenAnswer: string;
+  correct: boolean;
+  lessonPath: string;
+}
