+++ src/app/shared/models/topic.model.ts
@@ -1,0 +1,4 @@
+export interface TopicModel {
+  name: string;
+  path: string;
+}
